refactor(desktop): extract OmiHTTPTransport
Move URLSession setup, header assembly, GET/POST/DELETE verbs, 401 retry,
and JSON decode/error handling from APIClient into OmiHTTPTransport; APIClient
delegates with unchanged public surface. Add APIClient.swift line-count ratchet.

Verified: pytest test_app_client_swift_generator + test_desktop_rest_inventory;
xcrun swift test --filter APIClient (1 pre-existing routing assertion failure).
